test(manifest): drop the vocabulary and absence over-tests
The vocabulary "by name" pin and the absence shadowing trick were belt and
suspenders over what review, AGENTS.md and the behavioural suite already cover.
Delete both; keep the one unique test they held, insert-replaces, in
root_config.rs where the cross-format behaviour already lives.

AI Assistance: Claude Code (Claude Opus 4.8) made this change and its test move.
