+++ crates/integration-tests/Cargo.toml
@@ -126,10 +126,6 @@ name = "ldb_scan"
 path = "tests/ldb/scan.rs"
 
 [[test]]
-name = "manifest_absence"
-path = "tests/manifest/absence.rs"
-
-[[test]]
 name = "manifest_dyn_roundtrip"
 path = "tests/manifest/dyn_roundtrip.rs"
 
@@ -144,10 +140,6 @@ path = "tests/manifest/listing.rs"
 [[test]]
 name = "manifest_root_config"
 path = "tests/manifest/root_config.rs"
-
-[[test]]
-name = "manifest_vocabulary"
-path = "tests/manifest/vocabulary.rs"
 
 [[test]]
 name = "mantaray_bmt_nodes"
